fix: fix sword attack speed, balance update

+++ Assets/Scripts/Economy/WeaponUpgradeOption.cs
@@ -3,5 +3,5 @@ public class WeaponUpgradeOption : UpgradeOption
 {
     public WeaponBase targetWeapon;
     public bool isUnlock;
-    public int multiplier;
+    public float multiplier;
 }
